#1 removed animation function
removing the little animation function and swapping the image back out
to winged tyto makes more sense. the animation isn’t integral to how it
works so isn’t really necessary. although it would be good to remember
that triggering action events could be handy if tyto was embedded on a
page somewhere or the likes.